Get rid of more longs and all "%ld"s, new gcc + Windows + them = bad mojo. (cherry picked from commit 956b740cff3b29381f3c2255ac9523b507adcff9)

+++ crawl-ref/source/dbg-asrt.cc
@@ -156,8 +156,8 @@ static void _dump_player(FILE *file)
 
     if (you.delay_queue.size() > 0)
     {
-        fprintf(file, "Delayed (%lu):\n",
-                (unsigned long) you.delay_queue.size());
+        fprintf(file, "Delayed (%u):\n",
+                (unsigned int)you.delay_queue.size());
         for (unsigned int i = 0; i < you.delay_queue.size(); ++i)
         {
             const delay_queue_item &item = you.delay_queue[i];
