+++ crawl-ref/source/items.cc
@@ -2142,8 +2142,8 @@ static std::string _drop_selitem_text( const std::vector<MenuEntry*> *s )
         }
     }
 
-    return (make_stringf( " (%lu%s turn%s)",
-                (unsigned long) (s->size()),
+    return (make_stringf( " (%u%s turn%s)",
+                s->size(),
                 extraturns? "+" : "",
                 s->size() > 1? "s" : "" ));
 }
